% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absdiff.veloc.R
\name{absdiff.veloc}
\alias{absdiff.veloc}
\title{A diagnostic function for measurements of terminal velocity}
\usage{
absdiff.veloc(obj)
}
\arguments{
\item{obj}{The R object "vtobj.Rdata" produced by the script to calculate terminal velocity from .txt files}
}
\value{
Result summary of the absolute differences between observed and predicted velocity between successive images
}
\description{
To calculate absolute differences between observed and predicted velocity between successive images
}
\examples{
library(velocimeter)
load(file = paste0(.libPaths()[which("velocimeter"\%in\%list.files(.libPaths()))],"/velocimeter/extdata/vtobj.Rdata"))
absdiff.veloc(obj = vtobj[[1]])
}
