% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_terminal_velocity.R
\name{calculate.terminal.velocity.phys}
\alias{calculate.terminal.velocity.phys}
\title{A function to calculate terminal velocity}
\usage{
calculate.terminal.velocity.phys(
  file,
  min.size,
  min.circularity,
  fps = 130,
  tubelength
)
}
\arguments{
\item{file}{A .txt file containing pixel coordinates extracted from videos by imageJ}

\item{min.size}{Minimum size (in pixels) of valid objects}

\item{min.circularity}{Minimum circularity of valid objects}

\item{fps}{Frame per second, defaults to 130 in our method}

\item{tubelength}{2.095 m,1.575 m, and 1.075 m for long, medium, and short tube, respectively}
}
\value{
A list of estimated Vt using the linear and mechanistic models, data used to fit the models, intercept, R-squared values
}
\description{
The function allows one to calculate terminal velocity using a physical model for free fall with air resistance for a large sphere.
}
\examples{
library(velocimeter)
# not run
# calculate.terminal.velocity.phys(file = "terminal_velocity_analysis20219191543/agri-short_00000.aviResults.txt",min.size = 10,min.circularity = 0.05,fps = 130,tubelength = 1.075)
}
