% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{calculate.terminal.velocity.test}
\alias{calculate.terminal.velocity.test}
\title{A test function to calculate terminal velocity}
\usage{
calculate.terminal.velocity.test(
  file = "./inst/extdata/agri-short_00000.aviResults.txt"
)
}
\arguments{
\item{file}{A .txt file containing pixel coordinates extracted from videos by imageJ.
Here default to a file of a falling seed in Agrimonia eupatoria}
}
\value{
A list of estimated Vt using the linear and mechanistic models, data used to fit the models, intercept, R-squared values
}
\description{
Testing the function to calculate terminal velocity with a physical model for free fall with air resistance
}
