% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.veloc.R
\name{rmse.veloc}
\alias{rmse.veloc}
\title{A diagnostic function for measurements of terminal velocity}
\usage{
rmse.veloc(obj)
}
\arguments{
\item{obj}{The R object "vtobj.Rdata" produced by the script to calculate terminal velocity from .txt files}
}
\value{
Result summary of the RMSE of differences between observed and predicted velocity between successive images
}
\description{
To calculate the root-mean-square error (RMSE) of differences between observed and predicted velocity between successive images
}
\examples{
library(velocimeter)
load(file = paste0(.libPaths()[which("velocimeter"\%in\%list.files(.libPaths()))],"/velocimeter/extdata/vtobj.Rdata"))
rmse.veloc(obj = vtobj[[1]])
}
