% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcalib.R
\name{xcalib}
\alias{xcalib}
\title{Conversion function for real x coordinates of a falling object in 3-d space}
\usage{
xcalib(dat)
}
\arguments{
\item{dat}{A dataframe containing the image coordinates (ym, xd) and the real x coordinates}
}
\value{
Summary result of the fitted linear regression model.
}
\description{
The function converts the image coordinates (x coordinates in the drect view and
y coordinates in the mirror view) to real x coordinates in 3-d space
}
\examples{
datx <- read.csv(file = paste0(system.file(package = "velocimeter"),"/extdata/datx.csv"),header = TRUE)
xcalib(dat=datx)
}
