% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ycalib.R
\name{ycalib}
\alias{ycalib}
\title{Conversion function for real y coordinates of a falling object in 3-d space}
\usage{
ycalib(dat)
}
\arguments{
\item{dat}{A dataframe containing the image coordinates (ym, xd) and the real y coordinates (y.m)}
}
\value{
Summary result of the fitted linear regression model.
}
\description{
The function converts the image coordinates (x coordinates in the drect view and
y coordinates in the mirror view) to real y coordinates in 3-d space
}
\examples{
daty <- read.csv(file = paste0(system.file(package = "velocimeter"),"/extdata/daty.csv"),header = TRUE)
ycalib(dat=daty)
}
