% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcalib.R
\name{zcalib}
\alias{zcalib}
\title{Conversion function for real z coordinates of a falling object in 3-d space}
\usage{
zcalib(dat)
}
\arguments{
\item{dat}{A dataframe containing the image coordinates (zd, zm) and the real z coordinates (z.m)}
}
\value{
Summary result of the fitted linear regression model.
}
\description{
The function converts the image coordinates (z coordinates in the drect view and
z coordinates in the mirror view) to real z coordinates in 3-d space
}
\examples{
datz <- read.csv(file = paste0(system.file(package = "velocimeter"),"/extdata/datz.csv"),header = TRUE)
zcalib(dat=datz)
}
